drivebackup/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gspread==3.6.0
httplib2==0.18.1
idna==2.10
minio==7.0.1
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.3
